Add "Save as" dialog, and call it when we need a filename
Make LatexDocument.dir into a property.  When it's accessed, and it
isn't set, pop up a "Save as" dialog to get the directory and filename.
If the user doesn't choose a file, raise an ObstinateUserError.  Since
the fullfilename properties call dir, we can use the fullfilename and
know it is either set or an exception will be raised elsewhere.

In the future, we should probably make sure a filename is chosen on
loading, so we're not exposed to these problems.

+++ lib/latexslide.py
@@ -22,17 +22,15 @@ class LatexSlide(object):
     
     @property
     def fullfilename(self):
-        if self.parent.dir:
-            # Lazily create filename
-            if not self._filename:
-                # This is deprecated as a security risk, since the file could be
-                # created before you make it.  But we're not actually using this
-                # file; we're using the filename as a base, so the additional
-                # features of mkstemp don't help us avoid this problem.
-                filename = tempfile.mktemp(dir=self.parent.dir)
-                self._filename = os.path.basename(filename)
-            return os.path.join(self.parent.dir, self._filename)
-        return None
+        # Lazily create filename
+        if not self._filename:
+            # This is deprecated as a security risk, since the file could be
+            # created before you make it.  But we're not actually using this
+            # file; we're using the filename as a base, so the additional
+            # features of mkstemp don't help us avoid this problem.
+            filename = tempfile.mktemp(dir=self.parent.dir)
+            self._filename = os.path.basename(filename)
+        return os.path.join(self.parent.dir, self._filename)
     
     def set_content(self, content=""):
         self.buffer.begin_not_undoable_action()
@@ -70,5 +68,6 @@ class LatexSlide(object):
                 self.parent.pages[i][1] = self.pb
     
     def del_files(self):
-        for f in glob.glob(self.fullfilename + '*'):
-            os.unlink(f)
+        if self._filename:
+            for f in glob.glob(self.fullfilename + '*'):
+                os.unlink(f)
